requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
openpyxl>=3.1
click>=8.1
pyyaml>=6.0
jinja2>=3.1
requests>=2.31
dune-client>=1.7
python-dateutil>=2.8
tqdm>=4.65
python-dotenv>=1.0
